refactor: 카카오 로그인 로직 리팩터링 및 의존성 구조 개선
* refactor: 카카오 로그인 로직 core:data 모듈 이동 및 인터페이스화

- `feature:login:impl`에 위치하던 `KakaoLoginManager`를 `core:data` 모듈로 이동했습니다.
- `KakaoLoginManager`를 인터페이스로 정의하고, `KakaoLoginManagerImpl`에서 상세 로직을 구현하도록 변경했습니다.
- `KakaoLoginResult` sealed interface를 `core:data`로 이동했습니다.
- Hilt를 사용하여 `RepositoryModule`에서 `KakaoLoginManager` 의존성을 주입하도록 설정했습니다.

* refactor: LoginViewModel 내 카카오 로그인 비즈니스 로직 통합

- `LoginUiIntent`를 제거하고 `onClickLogin` 메서드로 직접 이벤트를 처리하도록 변경했습니다.
- `LoginScreen`에서 수행하던 카카오 로그인 호출 로직을 `LoginViewModel` 내부로 이동하여 뷰모델의 책임을 강화했습니다.
- `LoginUiEffect`에서 더 이상 사용되지 않는 `LaunchKakaoLogin`을 제거했습니다.
- 로그인 성공/실패 시의 상태 업데이트 및 Side Effect 처리 로직을 정리했습니다.

* build: 모듈 간 의존성 정리

- `core:data` 모듈에 `kakao.user` 라이브러리 의존성을 추가했습니다.
- `feature:login:impl`에서 `kakao.user` 의존성을 제거하고 `core:data`를 참조하도록 수정했습니다.
- `app` 모듈의 `build.gradle.kts`에서 불필요하게 명시되어 있던 `feature:login` 관련 의존성을 제거했습니다.

* style: LoginViewModel 및 LoginScreen 코드 스타일 정리

- `LoginViewModel` 및 `LoginScreen` 내 불필요한 공백을 제거하고 코드를 정렬했습니다.
- `LoginScreen`의 `onLogin` 콜백에서 Context와 실패 메시지를 ViewModel로 전달하도록 수정했습니다.
