+++ Prezel/app/build.gradle.kts
@@ -60,8 +60,6 @@ dependencies {
     implementation(projects.featureHistoryImpl)
     implementation(projects.featureProfileApi)
     implementation(projects.featureProfileImpl)
-    implementation(projects.featureLoginApi)
-    implementation(projects.featureLoginImpl)
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
