+++ Prezel/core/data/build.gradle.kts
@@ -12,4 +12,5 @@ dependencies {
     implementation(projects.coreNetwork)
 
     implementation(libs.kotlinx.coroutines.core)
+    implementation(libs.kakao.user)
 }
