+++ Prezel/core/data/src/main/java/com/team/prezel/core/data/auth/KakaoLoginManager.kt
@@ -1,0 +1,9 @@
+package com.team.prezel.core.data.auth
+
+import android.content.Context
+
+interface KakaoLoginManager {
+    suspend fun login(context: Context): KakaoLoginResult
+
+    suspend fun logout(): Result<Unit>
+}
