+++ Prezel/feature/login/impl/build.gradle.kts
@@ -7,7 +7,7 @@ android {
 }
 
 dependencies {
+    implementation(projects.coreData)
     implementation(projects.featureLoginApi)
     implementation(projects.featureHomeApi)
-    implementation(libs.kakao.user)
 }
