+++ Prezel/feature/login/impl/src/main/java/com/team/prezel/feature/login/impl/viewModel/LoginUiEffect.kt
@@ -1,8 +1,6 @@
 package com.team.prezel.feature.login.impl.viewModel
 
 sealed interface LoginUiEffect {
-    data object LaunchKakaoLogin : LoginUiEffect
-
     data object NavigateToHome : LoginUiEffect
 
     data class ShowSnackbar(
