--- Prezel/feature/login/impl/src/main/java/com/team/prezel/feature/login/impl/viewModel/LoginUiIntent.kt
@@ -1,5 +1,0 @@
-package com.team.prezel.feature.login.impl.viewModel
-
-sealed interface LoginUiIntent {
-    data object OnClickLogin : LoginUiIntent
-}
